# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medirush API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  owner: "owner",
} as const;

export interface User {
  id: string;
  fullName: string;
  phone: string;
  email: string;
  location: string;
  role: UserRole;
}

export interface SignupRequest {
  /** @minLength 2 */
  fullName: string;
  /** @minLength 8 */
  phone: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  location: string;
}

export interface LoginRequest {
  identifier: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Category {
  id: string;
  name: string;
}

export interface CategoryInput {
  /** @minLength 2 */
  name: string;
}

export interface Medicine {
  id: string;
  name: string;
  price: number;
  categoryId: string;
  categoryName: string;
  imageUrl: string;
  description: string;
}

export interface MedicineInput {
  /** @minLength 2 */
  name: string;
  /** @minimum 1 */
  price: number;
  categoryId: string;
  imageUrl: string;
  description: string;
}

export interface CartItem {
  medicine: Medicine;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
}

export interface AddCartItemRequest {
  medicineId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface UpdateCartItemRequest {
  /** @minimum 0 */
  quantity: number;
}

export interface PrescriptionInput {
  fileName: string;
  dataUrl: string;
}

export interface Prescription {
  id: string;
  fileName: string;
  imageUrl: string;
  createdAt: string;
}

export type CreateOrderRequestPaymentMethod =
  (typeof CreateOrderRequestPaymentMethod)[keyof typeof CreateOrderRequestPaymentMethod];

export const CreateOrderRequestPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export interface CreateOrderRequest {
  paymentMethod: CreateOrderRequestPaymentMethod;
  prescriptionId?: string;
  deliveryAddress: string;
}

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  upi: "upi",
} as const;

export interface Order {
  id: string;
  items: CartItem[];
  total: number;
  paymentMethod: OrderPaymentMethod;
  status: string;
  etaMinutes: number;
  prescriptionId?: string;
  deliveryAddress: string;
  createdAt: string;
  customerName?: string;
  customerPhone?: string;
}

export interface UpdateOrderStatusRequest {
  status: string;
}

export interface DashboardSummary {
  medicines: number;
  categories: number;
  orders: number;
  revenue: number;
}

export interface PaymentConfig {
  upiId: string;
  qrCodeImageUrl: string;
}

export type ListMedicinesParams = {
  search?: string;
  categoryId?: string;
};