# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Medirush API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: medicines
    description: Medicine catalogue operations
  - name: categories
    description: Category operations
  - name: cart
    description: Cart operations
  - name: orders
    description: Order operations
  - name: prescriptions
    description: Prescription upload operations
  - name: config
    description: App configuration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login as user or owner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /medicines:
    get:
      operationId: listMedicines
      tags: [medicines]
      summary: List medicines
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: categoryId
          schema:
            type: string
      responses:
        "200":
          description: Medicine list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Add medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Medicine created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    put:
      operationId: updateMedicine
      tags: [medicines]
      summary: Edit medicine
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "200":
          description: Medicine updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      summary: Delete medicine
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Medicine deleted
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Add category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Category deleted
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemRequest"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear cart
      responses:
        "200":
          description: Empty cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{medicineId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart quantity
      parameters:
        - in: path
          name: medicineId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemRequest"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove cart item
      parameters:
        - in: path
          name: medicineId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /prescriptions:
    post:
      operationId: uploadPrescription
      tags: [prescriptions]
      summary: Save prescription upload metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrescriptionInput"
      responses:
        "201":
          description: Prescription saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Checkout cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [orders]
      summary: Owner dashboard summary
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /config/payment:
    get:
      operationId: getPaymentConfig
      tags: [config]
      summary: Payment configuration
      responses:
        "200":
          description: Payment config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentConfig"
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        location:
          type: string
        role:
          type: string
          enum: [user, owner]
      required: [id, fullName, phone, email, location, role]
    SignupRequest:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        location:
          type: string
          minLength: 2
      required: [fullName, phone, email, password, location]
    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
        password:
          type: string
      required: [identifier, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required: [id, name]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
      required: [name]
    Medicine:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        categoryId:
          type: string
        categoryName:
          type: string
        imageUrl:
          type: string
        description:
          type: string
      required: [id, name, price, categoryId, categoryName, imageUrl, description]
    MedicineInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        price:
          type: number
          minimum: 1
        categoryId:
          type: string
        imageUrl:
          type: string
        description:
          type: string
      required: [name, price, categoryId, imageUrl, description]
    CartItem:
      type: object
      properties:
        medicine:
          $ref: "#/components/schemas/Medicine"
        quantity:
          type: integer
      required: [medicine, quantity]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
      required: [items, total]
    AddCartItemRequest:
      type: object
      properties:
        medicineId:
          type: string
        quantity:
          type: integer
          minimum: 1
      required: [medicineId, quantity]
    UpdateCartItemRequest:
      type: object
      properties:
        quantity:
          type: integer
          minimum: 0
      required: [quantity]
    PrescriptionInput:
      type: object
      properties:
        fileName:
          type: string
        dataUrl:
          type: string
      required: [fileName, dataUrl]
    Prescription:
      type: object
      properties:
        id:
          type: string
        fileName:
          type: string
        imageUrl:
          type: string
        createdAt:
          type: string
      required: [id, fileName, imageUrl, createdAt]
    CreateOrderRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          enum: [cod, upi]
        prescriptionId:
          type: string
        deliveryAddress:
          type: string
      required: [paymentMethod, deliveryAddress]
    Order:
      type: object
      properties:
        id:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cod, upi]
        status:
          type: string
        etaMinutes:
          type: integer
        prescriptionId:
          type: string
        deliveryAddress:
          type: string
        createdAt:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
      required: [id, items, total, paymentMethod, status, etaMinutes, deliveryAddress, createdAt]
    DashboardSummary:
      type: object
      properties:
        medicines:
          type: integer
        categories:
          type: integer
        orders:
          type: integer
        revenue:
          type: number
      required: [medicines, categories, orders, revenue]
    PaymentConfig:
      type: object
      properties:
        upiId:
          type: string
        qrCodeImageUrl:
          type: string
      required: [upiId, qrCodeImageUrl]