# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medirush API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from "./cartItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";

export interface Order {
  id: string;
  items: CartItem[];
  total: number;
  paymentMethod: OrderPaymentMethod;
  status: string;
  etaMinutes: number;
  prescriptionId?: string;
  deliveryAddress: string;
  createdAt: string;
  customerName?: string;
  customerPhone?: string;
}